services/cue-gateway/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.0
paho-mqtt==1.6.1
jsonschema==4.23.0
